% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighlyReplicatedRNASeq-package.R
\docType{package}
\name{HighlyReplicatedRNASeq}
\alias{HighlyReplicatedRNASeq}
\title{HighlyReplicatedRNASeq: Collection of Bulk RNA-Seq Experiments With Many Replicates}
\description{
The HighlyReplicatedRNASeq package provides access to the count matrix results
from studies with many replicates. These datasets can be valuable for
benchmarking tools designed to handle RNA-seq data.
}
\details{
Available datasets:
\itemize{
\item Schurch et al. (2016): 86 samples of S. \emph{cerevisiae} in two conditions
\itemize{
\item \code{\link[=Schurch16]{Schurch16()}} / \code{\link[=Schurch16_metadata]{Schurch16_metadata()}}
}
}

At the moment, this package contains only one dataset, but more
datasets can be added in the future.
}
